requirements.txt
certifi
click
idna
influxdb
python-dateutil
paho-mqtt
six
urllib3
pyyaml
